+++ frontend/src/calendar/Calendar.tsx
@@ -1,106 +1,69 @@
-import { useQuery } from "@apollo/client"
-import { lighten, makeStyles, Typography } from "@material-ui/core"
-import gql from "graphql-tag"
-import { Appointment } from "../types"
-import Day from "./Day"
+import { Calendar as BigCalendar, dateFnsLocalizer } from "react-big-calendar"
+import { format, parse, startOfWeek, getDay, parseISO } from "date-fns"
+import { de, enUS } from "date-fns/locale"
+import "react-big-calendar/lib/css/react-big-calendar.css"
+import { makeStyles } from "@material-ui/core"
+import useAppointmentAssistant from "./useAppointmentAssistant"
+import { useSnackbar } from "notistack"
+import { Loading } from "../util"
 
-interface Props {
-	date: Date
+
+const locales = {
+	"de": de,
+	"en-US": enUS
 }
 
-export default function Calendar(props: Props) {
+const localizer = dateFnsLocalizer({
+	format,
+	parse,
+	startOfWeek,
+	getDay,
+	locales
+})
+
+export default function Calendar() {
 	const classes = useStyles()
-	const { data } = useQuery<AppointmentsResult>(APPOINTMENTS)
-	const daysInMonth = new Date(props.date.getFullYear(), props.date.getMonth() + 1, 0).getDate()
-	const firstDayOfMonth = new Date(props.date.getFullYear(), props.date.getMonth(), 1)
-	const start = new Date(firstDayOfMonth.getFullYear(), firstDayOfMonth.getMonth(), firstDayOfMonth.getDate() - firstDayOfMonth.getDay() + 1)
+	const { enqueueSnackbar } = useSnackbar()
+	const { toggleAppointment, appointments, createAppointment, loading } = useAppointmentAssistant()
 
 	return (
 		<div className={classes.root}>
-			{
-				ROWS.map(rid =>
-					<div key={rid} className={classes.row}>
-						{
-							CELLS.map((dom, cid) =>
-								<div key={cid} className={classes.cell}>
-									<Typography color="textPrimary" variant="body2" align="center">
-										{rid === 0 ? <>{dom}<br /></> : null}
-										{new Date(start.getFullYear(), start.getMonth(), start.getDate() + rid * CELLS.length + cid).getDate()}
-									</Typography>
-									{
-										isSameDay(props.date, new Date(start.getFullYear(), start.getMonth(), start.getDate() + rid * CELLS.length + cid)) ?
-											<div className={classes.entry}>
-												Wizard!!!
-											</div> : null
-									}
-								</div>
-							)
+			<Loading loading={loading} />
+
+			<BigCalendar
+				localizer={localizer}
+				culture={navigator.language in locales ? navigator.language : "en-US"}
+				views={["week", "day", "agenda"]}
+				defaultView="week"
+				selectable={true}
+				onSelectSlot={async ({ start, end, action }) => {
+					if (action === "select") {
+						try {
+							await createAppointment(start as Date, end as Date)
+						} catch (err) {
+							console.error(err)
+							enqueueSnackbar("Event konnte nicht erstellt werden.", { variant: "error" })
 						}
-					</div>
-				)
-			}
-		</div >
+					}
+				}}
+				onDoubleClickEvent={async (appointment) => {
+					try {
+						await toggleAppointment(appointment)
+					} catch (err) {
+						console.error(err)
+						enqueueSnackbar("Aktion konnte nicht durchgeführt werden.", { variant: "error" })
+					}
+				}}
+				titleAccessor={appointment => appointment.participants.map(user => user.name).join(", ")}
+				startAccessor={appointment => parseISO(appointment.start)}
+				endAccessor={appointment => parseISO(appointment.end)}
+				events={appointments} />
+		</div>
 	)
 }
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles({
 	root: {
-		display: "flex",
-		flexDirection: "column",
-		height: "100%",
-		width: "100%"
-	},
-	row: {
-		display: "flex",
-		flexDirection: "row",
-		flexGrow: 1
-	},
-	cell: {
-		flexGrow: 1,
-		flexBasis: 0,
-		display: "flex",
-		flexDirection: "column",
-		borderRight: "solid 1px",
-		borderTop: "solid 1px",
-		borderRightColor: lighten(theme.palette.background.default, 0.5),
-		borderTopColor: lighten(theme.palette.background.default, 0.5),
-		"&:hover": {
-			backgroundColor: lighten(theme.palette.background.default, 0.2)
-		}
-	},
-	cellTitle: {
-		textAlign: "center"
-	},
-	entry: {
-		backgroundColor: theme.palette.primary.main,
-		borderRadius: 4,
-		margin: theme.spacing(1),
-		padding: theme.spacing(0.5),
-		textOverflow: ""
+		height: "100%"
 	}
-}))
-
-interface AppointmentsResult {
-	appointments: Appointment[]
-}
-
-const APPOINTMENTS = gql`
-	query {
-		appointments {
-			id
-			start
-			end
-			user {
-				id
-				name
-			}
-		}
-	}
-`
-
-function isSameDay(a: Date, b: Date) {
-	return !Math.floor(Math.abs(Date.UTC(a.getFullYear(), a.getMonth(), a.getDate()) - Date.UTC(b.getFullYear(), b.getMonth(), b.getDate())) / 1000 / 60 / 60 / 24)
-}
-
-const ROWS = [0, 1, 2, 3, 4, 5]
-const CELLS = ["MO", "DI", "MI", "DO", "FR", "SA", "SO"]
+})