+++ frontend/src/types.ts
@@ -66,5 +66,6 @@ export interface GameInfo {
 export interface Appointment {
 	id: number
 	start: string
+	end: string
 	participants: User[]
 }